Inline sized_interpolate_impl into interpolator_trait_impl
sized_interpolate_impl! was only called from interpolator_trait_impl!, so inline
the slice-to-array conversion logic (interpolate/interpolate_fast/batch_interpolate/
batch_interpolate_fast) directly into the trait impl macro. Removes the intermediate
macro layer while fully consolidating the trait impl generation.

55 lines saved net (-135 +80).

All tests passing (130 total), clippy clean.
