+++ src/interpolator/one/mod.rs
@@ -151,33 +151,7 @@ where
     batch_interpolate_impl!();
 }
 
-impl<D, S> Interpolator<D::Elem> for Interp1D<D, S>
-where
-    D: Data + RawDataClone + Clone,
-    D::Elem: Num + PartialOrd + Euclid + Copy + Debug,
-    S: Strategy1D<D> + Clone,
-{
-    /// Returns `1`.
-    #[inline]
-    fn ndim(&self) -> usize {
-        N
-    }
-
-    fn validate(&self) -> Result<(), ValidateError> {
-        self.check_extrapolate(&self.extrapolate)?;
-        self.data.validate()?;
-        self.validate_strategy()?;
-        Ok(())
-    }
-
-    sized_interpolate_impl!();
-
-    fn set_extrapolate(&mut self, extrapolate: Extrapolate<D::Elem>) -> Result<(), ValidateError> {
-        self.check_extrapolate(&extrapolate)?;
-        self.extrapolate = extrapolate;
-        Ok(())
-    }
-}
+interpolator_trait_impl!(Interp1D, Strategy1D, N);
 
 set_strategy_box_impl!(Interp1D, Strategy1D);
 set_strategy_enum_impl!(Interp1D, strategy::enums::Strategy1DEnum);
