+++ src/interpolator/two/mod.rs
@@ -165,33 +165,7 @@ where
     batch_interpolate_impl!();
 }
 
-impl<D, S> Interpolator<D::Elem> for Interp2D<D, S>
-where
-    D: Data + RawDataClone + Clone,
-    D::Elem: Num + PartialOrd + Euclid + Copy + Debug,
-    S: Strategy2D<D> + Clone,
-{
-    /// Returns `2`.
-    #[inline]
-    fn ndim(&self) -> usize {
-        N
-    }
-
-    fn validate(&self) -> Result<(), ValidateError> {
-        self.check_extrapolate(&self.extrapolate)?;
-        self.data.validate()?;
-        self.validate_strategy()?;
-        Ok(())
-    }
-
-    sized_interpolate_impl!();
-
-    fn set_extrapolate(&mut self, extrapolate: Extrapolate<D::Elem>) -> Result<(), ValidateError> {
-        self.check_extrapolate(&extrapolate)?;
-        self.extrapolate = extrapolate;
-        Ok(())
-    }
-}
+interpolator_trait_impl!(Interp2D, Strategy2D, N);
 
 set_strategy_box_impl!(Interp2D, Strategy2D);
 set_strategy_enum_impl!(Interp2D, strategy::enums::Strategy2DEnum);
